dom/tests/mochitest/bugs/test_bug684544.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <title>Test for Bug </title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug  **/

var f = document.createElement("iframe");
document.body.appendChild(f);
var win = f.contentWindow;

// Set location once to make sure it's resolved
win.location = "data:text/html,1";

// Now try to make the location object go away.
f.remove();

// Check that location is now null.  If it's not, the test needs changing
// (e.g. to use window.close() so that it's null).
is("location" in win, true, "Should still have a location property");
ok(!!win.location, "Location should not be null");

// Just set the location.  This should not crash but throw an exception.
var noException;
try {
  win.location = "data:text/html,2";
  noException = true;
}
catch(e) {
  noException = false;
}
todo(noException, "Shouldn't throw when setting location.");

// And check that we didn't override the slot in the process.
is(typeof(win.location), "object", "Location should not have become a string");
ok(!!win.location, "Location should not be null");

</script>
</pre>
</body>
</html>
